Change cursor on hovering the location link (#163)
On the page with detailed event description there is a block with links:
* link to the speaker's social network account
* link to the event in a social network (like vk meeting)
* link to map

But actually the last item isn't a link, it's just a button which triggers a modal window.
It also doesn't have an `href` attribute, so chrome doesn't change cursor type (to pointer).
This commit fixes this strange behaviour. Some people (including me) may be confused with it.

The better way to solve the problem is to handle requests like `it61.info/events/some-talk?map`.